src/app/components/controls/button/button.component.spec.ts
import { async, ComponentFixture, TestBed } from '@angular/core/testing';

import { ButtonComponent } from './button.component';

describe('ButtonComponent', (): void => {
  let component: ButtonComponent;
  let fixture: ComponentFixture<ButtonComponent>;

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      declarations: [ButtonComponent],
    }).compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(ButtonComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', (): void => {
    expect(component).toBeTruthy();
  });

  it('onClickButton() should raise buttonEvent', () => {
    component.disabled = false;

    component.buttonEvent.subscribe((credentials) => {
      expect(component.disabled).toBe(false);
      expect(credentials).toBe(credentials, 'send event');
    });

    component.onClickButton();
  });
});
